fix(refactor): template
