+++ src/components/FebCard.tsx
@@ -1,10 +1,11 @@
-import { Image } from '@/types/image';
-import React from 'react';
+import { CardActionArea } from '@mui/material';
 import Card from '@mui/material/Card';
 import CardContent from '@mui/material/CardContent';
 import CardMedia from '@mui/material/CardMedia';
 import Typography from '@mui/material/Typography';
-import { CardActionArea } from '@mui/material';
+import React from 'react';
+
+import { Image } from '@/types/image';
 
 interface FebCardProps {
   className?: string;
