+++ src/components/box-gradient.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { styled } from '@mui/material';
 import Box, { BoxProps } from '@mui/system/Box';
