+++ src/layouts/funny-layout/nav-content.tsx
@@ -1,7 +1,7 @@
-import React from 'react';
 import { Typography } from '@mui/material';
-import Grid from '@mui/material/Grid';
 import Box from '@mui/material/Box';
+import Grid from '@mui/material/Grid';
+import React from 'react';
 import type { FC } from 'react';
 
 interface NavContentProps {
