+++ src/layouts/main-layout/responsive-main-appbar/swipeable-slide-drawer.tsx
@@ -1,13 +1,14 @@
-import { NavMenuList } from '.';
-import * as React from 'react';
 import Box from '@mui/material/Box';
-import SwipeableDrawer from '@mui/material/SwipeableDrawer';
-import List from '@mui/material/List';
 import Divider from '@mui/material/Divider';
+import List from '@mui/material/List';
 import ListItem from '@mui/material/ListItem';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
+import SwipeableDrawer from '@mui/material/SwipeableDrawer';
 import router from 'next/router';
+import * as React from 'react';
+
+import { NavMenuList } from '.';
 
 type Anchor = 'top' | 'left' | 'bottom' | 'right';
