+++ src/lib/axios.ts
@@ -1,5 +1,6 @@
-import { API_URL } from '@/config';
 import Axios from 'axios';
+
+import { API_URL } from '@/config';
 
 export const axios = Axios.create({
   baseURL: API_URL,
