+++ src/pages/coworking-space/index.tsx
@@ -1,6 +1,7 @@
-import CoworkingSpaceView from '@/views/coworking-space-view';
-import React from 'react';
 import Head from 'next/head';
+import React from 'react';
+
+import CoworkingSpaceView from '@/views/coworking-space-view';
 
 const CoworkingSpace: React.FC = () => {
   return (
