+++ src/pages/team.tsx
@@ -1,5 +1,5 @@
-import React from 'react';
 import Head from 'next/head';
+import React from 'react';
 
 const Team: React.FC = () => {
   return (
