+++ src/views/home-view/index.tsx
@@ -1,7 +1,9 @@
+import React from 'react';
+
+import MainLayout from '@/layouts/main-layout';
+
 import { Hero } from './hero';
 import { SubScribe } from './sub-scribe';
-import MainLayout from '@/layouts/main-layout';
-import React from 'react';
 
 const HomeView: React.FC = () => {
   return (
