+++ theme/main-theme-provider.tsx
@@ -1,7 +1,8 @@
-import theme from './theme';
 import { CssBaseline } from '@mui/material';
 import { ThemeProvider } from '@mui/material/styles';
 import React from 'react';
+
+import theme from './theme';
 
 export const MainThemeProvider: React.FC = ({ children }) => {
   return (
